+++ eoreader/products/optical/vis1_product.py
@@ -658,3 +658,12 @@ class Vis1Product(VhrProduct):
             LOGGER.warning(f"No quicklook found in {self.condensed_name}")
 
         return quicklook_path
+
+    def _get_job_id(self) -> str:
+        """
+        Get VHR job ID
+
+        Returns:
+            str: VHR product ID
+        """
+        return self.split_name[-2]